pyproject.toml
[tool.poetry]
name = "shimoda-kr"
version = "0.1.0"
description = ""
authors = ["shimoshida <moshimoshimoshida@gmail.com>"]
readme = "README.md"
packages = [{include = "shimoda_kr"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.1.3"
matplotlib = "^3.9.2"
intervaltree = {git = "https://github.com/shimoshida/intervaltree.git", rev = "master"}
ipykernel = "^6.29.5"
dash = "^2.18.2"
dash-bootstrap-components = "^1.7.1"
pyyaml = "^6.0.2"


[[tool.poetry.source]]
name = "intervaltree"
url = "https://github.com/shimoshida/intervaltree.git"
priority = "explicit"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^25.1.0"
jupyter-black = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
